rename(ui-design-tool): split directories related to interaction
