+++ packages/ui-design-tool/components/InteractionController/utils/cursor.ts
