+++ packages/ui-design-tool/components/InteractionController/utils/rect.ts
